refactor: разделение сущностей, уход от OneToOne
