+++ common/src/main/java/com/github/jon7even/entity/user/UserEntity.java
@@ -1,20 +1,16 @@
 package com.github.jon7even.entity.user;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 import org.hibernate.annotations.ColumnDefault;
 
 import java.time.LocalDateTime;
@@ -61,10 +57,4 @@ public class UserEntity {
     @Column(name = "auth_on")
     @ColumnDefault(value = "false")
     private Boolean authorization;
-
-    @ToString.Exclude
-    @OneToOne(mappedBy = "user",
-            cascade = {CascadeType.PERSIST, CascadeType.REMOVE},
-            fetch = FetchType.LAZY, orphanRemoval = true)
-    private UserStatus userStatus;
 }
