+++ node/src/main/java/com/github/jon7even/entity/UserStatusEntity.java
@@ -1,22 +1,17 @@
-package com.github.jon7even.entity.user;
+package com.github.jon7even.entity;
 
 import com.github.jon7even.telegram.BotState;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.MapsId;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 
 /**
  * Класс описывающий статус пользователя.
@@ -31,7 +26,7 @@ import lombok.ToString;
 @AllArgsConstructor
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Table(name = "user_status", schema = "bot")
-public class UserStatus {
+public class UserStatusEntity {
 
     @Id
     @EqualsAndHashCode.Include
@@ -41,10 +36,4 @@ public class UserStatus {
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false)
     private BotState status;
-
-    @MapsId
-    @ToString.Exclude
-    @OneToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "chat_id", referencedColumnName = "chat_id", nullable = false)
-    private UserEntity user;
 }
